undef になる可能性があるので defined チェックする

+++ builder/MyBuilder.pm
@@ -7,7 +7,7 @@ use File::ShareDir ':ALL';
 sub ACTION_clean {
 	my $self = shift;
 	my $dir = eval { dist_dir('Cmd-Dwarf'); };
-	if (-d $dir) {
+	if (defined $dir and -d $dir) {
 		system("rm -rf $dir");
 	}
 	$self->SUPER::ACTION_clean;
@@ -16,7 +16,7 @@ sub ACTION_clean {
 sub ACTION_install {
 	my $self = shift;
 	my $dir = eval { dist_dir('Cmd-Dwarf'); };
-	if (-d $dir) {
+	if (defined $dir and -d $dir) {
 		system("rm -rf $dir");
 	}
 	$self->SUPER::ACTION_install;
